+++ logtape/mod.ts
@@ -41,7 +41,7 @@ export {
   type LogLevel,
   parseLogLevel,
 } from "./level.ts";
-export { getLogger, type Logger } from "./logger.ts";
+export { getLogger, type Logger, type LogMethod } from "./logger.ts";
 export type { LogRecord } from "./record.ts";
 export {
   type ConsoleSinkOptions,